+++ R/plot.R
@@ -129,12 +129,13 @@ theme_custom_map <- function(base_size = 11,
 #' @importFrom  rlang sym enquo
 #' @importFrom classInt classIntervals
 thematic_map <- function(
-    zoom = c("mega", "cty", "admi")[1], subset = NULL, stat = NULL, 
+    zoom = c("mega", "cty", "admi"), subset = NULL, stat = NULL, 
     polygon = TRUE, point = FALSE, label = NULL, col_cnt = 9, palette = "YlOrRd", 
     line_col = "darkgray", fill = "lightblue", point_col = "blue", title = NULL, 
     subtitle = NULL, legend_pos = c("none", "right", "left", "bottom", "top"),
     base_family = "NanumSquare")
 {
+  zoom <- match.arg(zoom)
   legend_pos <- match.arg(legend_pos)
   
   map_df <- eval(get(zoom))